Changes in support of plot set 11

+++ vistrails/gui/uvcdat/diagnosticsDockWidget.py
@@ -12,6 +12,7 @@ import metrics.frontend.defines as defines
 import metrics.fileio.findfiles
 import metrics.packages.diagnostic_groups
 from metrics.common.utilities import natural_sort
+import pdb
 
 class DiagnosticsDockWidget(QtGui.QDockWidget, Ui_DiagnosticDockWidget):
    dg_menu = metrics.packages.diagnostic_groups.diagnostics_menu()  # typical item: 'AMWG':AMWG
@@ -768,21 +769,34 @@ class DiagnosticsDockWidget(QtGui.QDockWidget, Ui_DiagnosticDockWidget):
             mbox = QtGui.QMessageBox(QtGui.QMessageBox.Warning, msg, QString(msg))
             mbox.exec_()
         ires = 0
+        #pdb.set_trace()
         for row in range(Nrows):
             for col in range(Ncolumns):
                if ires<len(res):
                   res30 = res[ires]
                else:
                   res30 = None
-               self.displayCell(res30, row, col)
+               
+               if type(res30) is not tuple:
+                   res30 = (res30,)
+               for plot_number, RES30 in enumerate(res30):
+                   self.displayCell(RES30, row, col, dropInfo=(plot_number == 0))
+
                ires += 1
 
-   def displayCell(self, res30, row, column, sheet="Sheet 1"):
+   def displayCell(self, res30, row, column, sheet="Sheet 1", dropInfo = True):
       """Display result into one cell defined by row/col args"""
+      #pdb.set_trace()
+      #print 'displayCell start'
+      #print dropInfo
+      #print "row, column =", row, column
       projectController = self.parent().get_current_project_controller()
-      projectController.get_sheet_widget(sheet).deleteCell(row,column)
+      if dropInfo:
+          projectController.get_sheet_widget(sheet).deleteCell(row,column)
       projectController.enable_animation = False  # I (JfP) don't know why I need this, it didn't
                                                   # used to be necessary.
+
+      #print 'res30=', res30
       if res30 is None:
          return
       if not hasattr(res30,'presentation') or res30.presentation is None or res30.presentation is "text":
@@ -803,14 +817,20 @@ class DiagnosticsDockWidget(QtGui.QDockWidget, Ui_DiagnosticDockWidget):
       import cdms2
       from packages.uvcdat_cdms.init import CDMSVariable
       from core.utils import InstanceObject
-
       from metrics.frontend.uvcdat import diagnostics_template
+      #pdb.set_trace()
       tm = diagnostics_template()  # template name is 'diagnostic'
+      if dropInfo:
+          tmplDropInfo = ('diagnostic', sheet, row, column)
+          projectController.template_was_dropped(tmplDropInfo)
       for V in pvars:
+         #print V.id, V
+         #pdb.set_trace()
          V.title = title        # VCS looks the title of the variable, not the plot.
          V.long_name = V.title  # VCS overrides title with long_name!
-         tmplDropInfo = ('diagnostic', sheet, row, column)
-         projectController.template_was_dropped(tmplDropInfo)
+         #tmplDropInfo = ('diagnostic', sheet, row, column)
+         #if dropInfo:
+         #    projectController.template_was_dropped(tmplDropInfo)
          # Until I know better storing vars in tempfile....
          f = tempfile.NamedTemporaryFile()
          filename = f.name
@@ -839,42 +859,45 @@ class DiagnosticsDockWidget(QtGui.QDockWidget, Ui_DiagnosticDockWidget):
          projectController.add_defined_variable(cdmsVar)
          
          # simulate drop variable
+         #print "SI DROPPING:",name_in_var_widget,row,column
          varDropInfo = (name_in_var_widget, sheet, row, column)
          projectController.variable_was_dropped(varDropInfo)
-         # Trying to add method to plot list....
-         #from gui.application import get_vistrails_application
-         #_app = get_vistrails_application()
-         #d = _app.uvcdatWindow.dockPlot
-         # simulate drop plot
-         pm = projectController.plot_manager
-         V=pm._plot_list["VCS"]
-         gm = res30.presentation
-         from packages.uvcdat_cdms.init import get_canvas, get_gm_attributes, original_gm_attributes
-         from gui.uvcdat.uvcdatCommons import gmInfos
-         Gtype = res30.type
-         if Gtype == "Taylor":
-             Gtype = "Taylordiagram"
+         #pdb.set_trace()
+      # Trying to add method to plot list....
+      #from gui.application import get_vistrails_application
+      #_app = get_vistrails_application()
+      #d = _app.uvcdatWindow.dockPlot
+      # simulate drop plot
+      pm = projectController.plot_manager
+      VCS_LIST = pm._plot_list["VCS"]
+      gm = res30.presentation
+      from packages.uvcdat_cdms.init import get_canvas, get_gm_attributes, original_gm_attributes
+      from gui.uvcdat.uvcdatCommons import gmInfos
+      Gtype = res30.type
+      if Gtype == "Taylor":
+          Gtype = "Taylordiagram"
+ 
+      G = VCS_LIST[Gtype]
+      if not gm.name in G.keys():
+         G[gm.name] = pm._registry.add_plot(gm.name,"VCS",None,None,Gtype)
+         G[gm.name].varnum = int(gmInfos[Gtype]["nSlabs"])
+ 
+      #add initial attributes to global dict
+      canvas = get_canvas()
+      method_name = "get"+Gtype.lower()
+      attributes = get_gm_attributes(Gtype)
 
-         G = V[Gtype]
-         if not gm.name in G.keys():
-            G[gm.name] = pm._registry.add_plot(gm.name,"VCS",None,None,Gtype)
-            G[gm.name].varnum = int(gmInfos[Gtype]["nSlabs"])
+      attrs = {}
+      for attr in attributes:
+         attrs[attr] = getattr(gm,attr)
+      original_gm_attributes[Gtype][gm.name] = InstanceObject(**attrs)
 
-         #add initial attributes to global dict
-         canvas = get_canvas()
-         method_name = "get"+Gtype.lower()
-         attributes = get_gm_attributes(Gtype)
-
-         attrs = {}
-         for attr in attributes:
-            attrs[attr] = getattr(gm,attr)
-         original_gm_attributes[Gtype][gm.name] = InstanceObject(**attrs)
-
-         plot = projectController.plot_manager.new_plot('VCS', Gtype, gm.name )
-         #plot = projectController.plot_manager.new_plot('VCS', Gtype, "default" )
-         plotDropInfo = (plot, sheet, row, column)
-         projectController.plot_was_dropped(plotDropInfo)
-
+      #print "PLOTTING:",Gtype,gm.name
+      plot = projectController.plot_manager.new_plot('VCS', Gtype, gm.name )
+      #plot = projectController.plot_manager.new_plot('VCS', Gtype, "default" )
+      plotDropInfo = (plot, sheet, row, column)
+      #pdb.set_trace()
+      projectController.plot_was_dropped(plotDropInfo)
    def cancelClicked(self):
         self.close()
